Add beginning/end-of-defun
* hare-mode.el
(hare-mode--regexp-declaration-line-beginning): new regexp matching keywords
that are likely to begin a line, as well as a code block.

(hare-mode--regexp-declaration-end): new regexp matching declarations ending
code blocks.

(hare-mode-beginning-of-defun): search for the previous function or other major
code block starter symbol.  Add this to the local binding.

(hare-mode-end-of-defun): search for the next function or other major code block
starter symbol.  Add this to the local binding.