Find tokens and return offsets
* hare-mode.el: two functions, 'hare-mode--find-token' and
'hare-mode--indent-offset-from-token' are helper functions to return what offset
we want to indent to.  The first one finds it on the same line (bad idea?), and
returns t or nil.  The latter returns the cons cell '(column . n)', where 'n' is
the wanted offset.  We can choose extra levels of offset, but defaulting to 1.
I definitely think there are better ways to do this, and am a little concerned
with adding too much logic, since I'm not quite sure the grammar is as good as
it needs to be.